WIP Module loader and parser for new config
